Extract useFetchMovieCharacters custom hooks into a new file
